+++ java-api/src/main/java/com/sap/cloud/security/token/Token.java
@@ -17,10 +17,13 @@ import java.security.Principal;
 import java.time.Instant;
 import java.util.*;
 
+import static com.sap.cloud.security.token.TokenClaims.ISSUER;
+
 /**
  * Represents a JSON Web Token (JWT).
  */
 public interface Token extends Serializable {
+
 	@SuppressWarnings("unchecked")
 	List<TokenFactory> services = new ArrayList() {
 		{
@@ -185,6 +188,19 @@ public interface Token extends Serializable {
 	 * @return the OAuth client ID.
 	 */
 	String getClientId();
+
+	/**
+	 * Returns the identifier for the Issuer of the token.
+	 * Its a URL that contains scheme, host, and optionally, port number and path components
+	 * but no query or fragment components.
+	 * This one is validated in the {@code JwtIssuerValidator} and
+	 * used as base url to discover jwks_uri endpoint for downloading the token keys.
+	 *
+	 * @return the issuer.
+	 */
+	default String getIssuer() {
+		return getClaimAsString(ISSUER);
+	}
 
 	/**
 	 * Returns the grant type of the jwt token. <br>