+++ java-security/src/test/java/com/sap/cloud/security/token/validation/validators/JwtIssuerValidatorTest.java
@@ -5,7 +5,9 @@
  */
 package com.sap.cloud.security.token.validation.validators;
 
+import com.sap.cloud.security.token.SapIdToken;
 import com.sap.cloud.security.token.Token;
+import com.sap.cloud.security.token.TokenClaims;
 import com.sap.cloud.security.token.validation.ValidationResult;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -19,12 +21,9 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 
-import static com.sap.cloud.security.token.TokenClaims.IAS_ISSUER;
-import static com.sap.cloud.security.token.TokenClaims.ISSUER;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.hamcrest.CoreMatchers.*;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.when;
 
 class JwtIssuerValidatorTest {
@@ -36,7 +35,7 @@ class JwtIssuerValidatorTest {
 	@BeforeEach
 	void setup() {
 		cut = new JwtIssuerValidator(Arrays.asList(domains));
-		token = Mockito.mock(Token.class);
+		token = Mockito.mock(SapIdToken.class);
 	}
 
 	@Test
@@ -50,47 +49,47 @@ class JwtIssuerValidatorTest {
 
 	@Test
 	void validationFails_whenIssuerDomainDoesNotMatchIdentityProviderDomains() {
-		when(token.getClaimAsString(ISSUER)).thenReturn("https://otherdomain.test.ondemand.com");
+		configureMock("https://otherdomain.test.ondemand.com", null);
 		assertThat(cut.validate(token).isValid(), is(false));
 	}
 
 	@ParameterizedTest
 	@NullAndEmptySource
 	@ValueSource(strings = { "  " })
-	void validationFails_whenIssuerIsEmptyOrNull(String issuer) {
+	void validationIgnoresEmptyIssuer_whenIasIssuerIsGiven(String issuer) {
 		cut = new JwtIssuerValidator(Collections.singletonList("accounts400.ondemand.com"));
-		token = Mockito.mock(Token.class);
+		configureMock(issuer, "https://test.accounts400.ondemand.com");
 
-		when(token.getClaimAsString(ISSUER)).thenReturn(issuer);
-		when(token.getClaimAsString(IAS_ISSUER)).thenReturn("https://test.accounts400.ondemand.com");
 		ValidationResult validationResult = cut.validate(token);
-		assertThat(validationResult.isErroneous(), is(true));
-		assertThat(validationResult.getErrorDescription(),
-				startsWith("Issuer validation can not be performed because Jwt token does not contain 'iss' claim."));
+		assertThat(validationResult.isValid(), is(true));
 	}
 
-	@ParameterizedTest
-	@NullAndEmptySource
-	@ValueSource(strings = { "  " })
-	void validationSucceeds_whenIasIssuerIsEmptyOrNull(String iasIssuer) {
+	@Test
+	void validationSucceeds_whenIasIssuerIsEmptyOrNull() {
 		cut = new JwtIssuerValidator(Collections.singletonList("accounts400.ondemand.com"));
-		token = Mockito.mock(Token.class);
+		configureMock("https://test.accounts400.ondemand.com", null);
 
-		when(token.getClaimAsString(ISSUER)).thenReturn("https://test.accounts400.ondemand.com");
-		when(token.getClaimAsString(IAS_ISSUER)).thenReturn(iasIssuer);
 		ValidationResult validationResult = cut.validate(token);
-		assertTrue(validationResult.isValid());
+		assertThat(validationResult.isErroneous(), is(false));
 	}
 
 	@Test
 	void validationFails_withoutMatchingIasIssuer() {
-		when(token.getClaimAsString(ISSUER)).thenReturn("https://otherDomain.accounts.ondemand.com");
-		when(token.getClaimAsString(IAS_ISSUER)).thenReturn("https://iasDomain.accounts.ondemand.com");
+		configureMock("https://otherDomain.accounts400.ondemand.com", "https://iasDomain.accounts.ondemand.com");
 
 		ValidationResult validationResult = cut.validate(token);
-		assertTrue(validationResult.isErroneous());
+		assertThat(validationResult.isErroneous(), is(true));
 		assertThat(validationResult.getErrorDescription(), startsWith(
-				"Issuer is not trusted because 'ias_iss' 'https://iasDomain.accounts.ondemand.com' doesn't match any of these domains '[customer.ondemand.com, accounts400.ondemand.com]' of the identity provider."));
+				"Issuer is not trusted because issuer claim 'https://iasDomain.accounts.ondemand.com' doesn't match any of these domains '[customer.ondemand.com, accounts400.ondemand.com]' of the identity provider."));
+	}
+
+	@Test
+	void validationIgnoresInvalidIssuer_whenIasIssuerIsGiven() {
+		cut = new JwtIssuerValidator(Arrays.asList(domains));
+		configureMock("invalid_url", "https://otherDomain.accounts400.ondemand.com");
+
+		ValidationResult validationResult = cut.validate(token);
+		assertThat(validationResult.isErroneous(), is(false));
 	}
 
 	@ParameterizedTest
@@ -99,34 +98,13 @@ class JwtIssuerValidatorTest {
 			"\0://myauth.com",
 			"https://otherDomain.org?accounts400.ondemand.com",
 			"subdomain.accounts400.ondemand.com" })
-	void validationFails_issuerUrl(String issuer) {
+	void validationFails_iasIssuerUrl(String iasIssuer) {
 		cut = new JwtIssuerValidator(Arrays.asList(domains));
-		token = Mockito.mock(Token.class);
-
-		when(token.getClaimAsString(ISSUER)).thenReturn(issuer);
-		when(token.getClaimAsString(IAS_ISSUER)).thenReturn("https://otherDomain.accounts400.ondemand.com");
+		configureMock("https://otherDomain.accounts400.ondemand.com", iasIssuer);
 
 		ValidationResult validationResult = cut.validate(token);
-		assertTrue(validationResult.isErroneous());
-		assertThat(validationResult.getErrorDescription(), startsWith("Issuer is not trusted because 'iss' claim"));
-	}
-
-	@ParameterizedTest
-	@CsvSource({ "https://subdomain.accounts400.ondemand.com#anyFragment_keys",
-			"https://subdomain.accounts400.ondemand.com?a=b",
-			"\0://myauth.com",
-			"https://otherDomain.org?accounts400.ondemand.com",
-			"subdomain.accounts400.ondemand.com" })
-	void validationFails_iasIssuerUrl(String issuer) {
-		cut = new JwtIssuerValidator(Arrays.asList(domains));
-		token = Mockito.mock(Token.class);
-
-		when(token.getClaimAsString(ISSUER)).thenReturn("https://otherDomain.accounts400.ondemand.com");
-		when(token.getClaimAsString(IAS_ISSUER)).thenReturn(issuer);
-
-		ValidationResult validationResult = cut.validate(token);
-		assertTrue(validationResult.isErroneous());
-		assertThat(validationResult.getErrorDescription(), startsWith("Issuer is not trusted because 'ias_iss' claim"));
+		assertThat(validationResult.isErroneous(), is(true));
+		assertThat(validationResult.getErrorDescription(), startsWith("Issuer is not trusted because issuer claim"));
 	}
 
 	@ParameterizedTest
@@ -135,12 +113,16 @@ class JwtIssuerValidatorTest {
 			"https://nestle.com,https://paas.accounts400.ondemand.com," })
 	void validationSucceeds(String issuer, String iasIssuer) {
 		cut = new JwtIssuerValidator(Arrays.asList(domains));
-		token = Mockito.mock(Token.class);
-
-		when(token.getClaimAsString(ISSUER)).thenReturn(issuer);
-		when(token.getClaimAsString(IAS_ISSUER)).thenReturn(iasIssuer);
+		configureMock(issuer, iasIssuer);
 
 		ValidationResult validationResult = cut.validate(token);
-		assertTrue(validationResult.isValid());
+		assertThat(validationResult.isValid(), is(true));
+	}
+
+	private void configureMock(String issuer, String iasIssuer) {
+		when(token.getIssuer()).thenCallRealMethod();
+		when(token.getClaimAsString(TokenClaims.ISSUER)).thenReturn(issuer);
+		when(token.getClaimAsString(TokenClaims.IAS_ISSUER)).thenReturn(iasIssuer);
+		when(token.hasClaim("ias_iss")).thenReturn(iasIssuer != null);
 	}
 }